archives/legacy_services/blog-launchpad/requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.7
requests==2.31.0
python-dotenv==1.0.0
humanize==4.8.0
pytz==2023.3
facebook-sdk==3.1.0 